move OpenAccount and Password to OS.Account
